+++ packages/orq-rc/RELEASES.md
@@ -689,3 +689,13 @@ Based on:
 - [typescript v3.2.0-rc.29] packages/orq-rc
 ### Releases
 - [NPM v3.2.0-rc.29] https://www.npmjs.com/package/@orq-ai/node/v/3.2.0-rc.29 - packages/orq-rc
+
+## 2025-03-06 18:16:09
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.512.0 (2.544.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.2.0-rc.30] packages/orq-rc
+### Releases
+- [NPM v3.2.0-rc.30] https://www.npmjs.com/package/@orq-ai/node/v/3.2.0-rc.30 - packages/orq-rc